move acl parcel to common
